--- tasks/run_cd4pe_job.json
@@ -1,41 +1,0 @@
-{
-  "description": "Retrieves the Puppet control repo and job script for the specified job instance id.",
-  "parameters": {
-    "job_instance_id": {
-      "type": "String[1]",
-      "description": "Instance ID of job."
-    },
-    "cd4pe_web_ui_endpoint": {
-      "type": "String[1]",
-      "description": "web ui endpoint of CD4PE. Used to download control repo and job scripts to execute jobs."
-    },
-    "cd4pe_token": {
-      "type": "String[1]",
-      "description": "Token used to hit Tokened CD4PE endpoints."
-    },
-    "cd4pe_job_owner": {
-      "type": "String[1]",
-      "description": "Account (workspace) owner of the job to be run."
-    },
-    "env_vars": {
-      "type": "Optional[Array[String[1]]]",
-      "description": "Environment variables to be set on the agent running the job."
-    },
-    "docker_image": {
-      "type": "Optional[String[1]]",
-      "description": "If specified, te job will attempt to run inside the docker container."
-    },
-    "docker_run_args": {
-      "type": "Optional[Array[String[1]]]",
-      "description": "If specified, the arguements will be passed to docker if docker image is specified."
-    }
-  },
-  "files": [
-    "cd4pe_deployments/lib/puppet_x/puppetlabs/cd4pe_client.rb",
-    "cd4pe_deployments/lib/puppet_x/puppetlabs/gzip_helper.rb",
-    "cd4pe_deployments/lib/puppet_x/puppetlabs/cd4pe_job_helper.rb",
-    "cd4pe_deployments/lib/puppet_x/puppetlabs/logger.rb"
-  ],
-
-  "input_method": "stdin"
-}
